feat : 닉네임 한줄평 조회 기능 구현
