+++ src/main/java/com/ureca/picky_be/base/business/lineReview/LineReviewUseCase.java
@@ -1,6 +1,7 @@
 package com.ureca.picky_be.base.business.lineReview;
 
 import com.ureca.picky_be.base.business.lineReview.dto.*;
+import com.ureca.picky_be.base.business.user.dto.UserLineReviewsReq;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
 
@@ -13,5 +14,5 @@ public interface LineReviewUseCase {
     UpdateLineReviewResp updateLineReview(Long lineReviewId, UpdateLineReviewReq req);
 
     Slice<ReadLineReviewResp> getLineReviewsByMovie(PageRequest pageRequest, LineReviewQueryRequest queryReq);
-
+    Slice<ReadLineReviewResp> getLineReviewsByNickname(PageRequest pageRequest, UserLineReviewsReq req);
 }
