+++ src/main/java/com/ureca/picky_be/base/business/user/UserService.java
@@ -66,11 +66,5 @@ public class UserService implements UserUseCase {
         return userDtoMapper.toGetNicknameValidationResp(userManager.getNicknameValidation(nickname));
     }
 
-    @Override
-    @Transactional(readOnly = true)
-    public Slice<ReadLineReviewResp> getLineReviewsByNickname(PageRequest pageRequest, UserLineReviewsReq req) {
-        Long userId = userManager.getUserIdByNickname(req.nickname());
-        Slice<LineReviewProjection> lineReviews = lineReviewManager.findLineReviewsByNickname(userId, req, pageRequest);
-        return lineReviews.map(lineReviewDtoMapper::toReadLineReviewResp);
-    }
+
 }
