+++ src/main/java/com/ureca/picky_be/base/persistence/lineReview/LineReviewRepository.java
@@ -67,7 +67,7 @@ public interface LineReviewRepository extends JpaRepository<LineReview, Long> {
                lr.createdAt AS createdAt
         FROM LineReview lr
         LEFT JOIN LineReviewLike lrl ON lrl.lineReview.id = lr.id
-        WHERE lr.userId = :userId AND lr.id < :lastReviewId
+        WHERE lr.userId = :userId AND lr.id > :lastReviewId
         GROUP BY lr.id, lr.userId, lr.movieId, lr.rating, lr.context, lr.isSpoiler, lr.createdAt
         ORDER BY lr.createdAt DESC
 """)
